pyproject.toml
[project]
name = "rag-source"
version = "0.1.0"
description = "The definitive guide to parsing unstructured documents for RAG"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # PDF
    "pypdf>=4.0",
    "pdfplumber>=0.11",
    "pymupdf>=1.24",
    "reportlab>=4.0",
    # DOCX
    "python-docx>=1.1",
    "mammoth>=1.8",
    "docx2txt>=0.8",
    # PPTX
    "python-pptx>=1.0",
    # HTML
    "beautifulsoup4>=4.12",
    "html2text>=2024.2",
    "trafilatura>=1.12",
    "lxml>=5.0",
    # Spreadsheets
    "openpyxl>=3.1",
    "pandas>=2.2",
    # Markdown
    "mistune>=3.0",
    # EPUB
    "ebooklib>=0.18",
    # Images
    "Pillow>=10.0",
    # Utilities
    "tabulate>=0.9",
]

[project.optional-dependencies]
ocr = [
    "pytesseract>=0.3",
    "easyocr>=1.7",
    "pdf2image>=1.17",
]
